[BAEL-7824] - Using Mapstruct with Inheritance.  * Changing folder structure.
