--- mapstruct/src/main/java/com/baeldung/inheritance/mapper/CarMapper.java
@@ -1,12 +1,0 @@
-package com.baeldung.inheritance.mapper;
-
-import org.mapstruct.Mapper;
-
-import com.baeldung.inheritance.model.Car;
-import com.baeldung.inheritance.model.dto.CarDTO;
-
-@Mapper()
-public interface CarMapper {
-
-    CarDTO carToDTO(Car car);
-}
