--- mapstruct/src/main/java/com/baeldung/inheritance/model/Visitor.java
@@ -1,10 +1,0 @@
-package com.baeldung.inheritance.model;
-
-import com.baeldung.inheritance.model.dto.VehicleDTO;
-
-public interface Visitor {
-
-    VehicleDTO visit(Car car);
-
-    VehicleDTO visit(Bus bus);
-}
