+++ mapstruct/src/main/java/com/baeldung/mapstruct/inheritance/mapper/BusMapper.java
@@ -1,0 +1,12 @@
+package com.baeldung.mapstruct.inheritance.mapper;
+
+import org.mapstruct.Mapper;
+
+import com.baeldung.mapstruct.inheritance.model.Bus;
+import com.baeldung.mapstruct.inheritance.model.dto.BusDTO;
+
+@Mapper()
+public interface BusMapper {
+
+    BusDTO busToDTO(Bus bus);
+}
