+++ mapstruct/src/main/java/com/baeldung/mapstruct/inheritance/mapper/CarMapper.java
@@ -1,0 +1,12 @@
+package com.baeldung.mapstruct.inheritance.mapper;
+
+import org.mapstruct.Mapper;
+
+import com.baeldung.mapstruct.inheritance.model.Car;
+import com.baeldung.mapstruct.inheritance.model.dto.CarDTO;
+
+@Mapper()
+public interface CarMapper {
+
+    CarDTO carToDTO(Car car);
+}
