+++ mapstruct/src/main/java/com/baeldung/mapstruct/inheritance/model/Visitor.java
@@ -1,0 +1,10 @@
+package com.baeldung.mapstruct.inheritance.model;
+
+import com.baeldung.mapstruct.inheritance.model.dto.VehicleDTO;
+
+public interface Visitor {
+
+    VehicleDTO visit(Car car);
+
+    VehicleDTO visit(Bus bus);
+}
