+++ mapstruct/src/main/java/com/baeldung/mapstruct/inheritance/model/dto/BusDTO.java
@@ -1,4 +1,4 @@
-package com.baeldung.inheritance.model.dto;
+package com.baeldung.mapstruct.inheritance.model.dto;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
